+++ yast/middlewares/__init__.py
@@ -3,6 +3,7 @@ __all__ = [
     "CORSMiddleware",
     "GZipMiddleware",
     "HttpsRedirectMiddleware",
+    "LifespanMiddleware",
     "TrustedHostMiddleware",
     "WSGIMiddleware",
     "BaseHttpMiddleware",
@@ -16,6 +17,7 @@ from .errors import ServerErrorMiddleware
 from .exception import ExceptionMiddleware
 from .gzip import GZipMiddleware
 from .httpsredirect import HttpsRedirectMiddleware
+from .lifespan import LifespanMiddleware
 from .sessions import SessionMiddleware
 from .trustedhost import TrustedHostMiddleware
 from .wsgi import WSGIMiddleware